+++ CHANGES.md
@@ -1,6 +1,6 @@
 
 
-- Add `At.role`.
+- Add `At.{role,popover,popoveraction,popovertargetaction`}.
 
 v0.1.0 2023-07-30 Zagreb
 ------------------------